Apply skeletal implementation to interface migration automated refactoring.
* Applied skeletal implementation to interface migration automated refactoring.

Signed-off-by: Raffi Khatchadourian <rkhatchadourian@citytech.cuny.edu>

* Unmigrate methods that crossed module boundaries.

Signed-off-by: Raffi Khatchadourian <rkhatchadourian@citytech.cuny.edu>

* Convert tab characters to spaces.

Signed-off-by: Raffi Khatchadourian <rkhatchadourian@citytech.cuny.edu>

* Remove unused import straggler.

Signed-off-by: Raffi Khatchadourian <rkhatchadourian@citytech.cuny.edu>

* Move curly brace to newline to conform to destination interface coding styles.

There are differences in coding styles between the source module, where the implementations originated, and the destination module, where the default methods were created. This commit is to conform to to new coding style. This is a manual step as we do not currently analyze the difference in coding styles between projects/modules.

Signed-off-by: Raffi Khatchadourian <rkhatchadourian@citytech.cuny.edu>

* Removed skeletal implementations from template file.

These implementations were automatically copied to PrimitiveIterable. This step represents the corresponding removable. This is a manual step and a workaround of https://github.com/khatchad/Migrate-Skeletal-Implementation-to-Interface-Refactoring/issues/165.

Signed-off-by: Raffi Khatchadourian <rkhatchadourian@citytech.cuny.edu>