Substantially improve efficiency of size method for concatenated rich iterables.
Override size method in CompositeIterable to call Iterate::sizeOf, which has a
special-case optimization when its parameter is a RichIterable.
This improves the complexity of the size method of concatenated rich iterables
from O(n), where n is the total number of elements in the concatenation, to
O(k), where k is the number of concatenated rich iterables.

This fixes #279.

Signed-off-by: Olivier Cinquin <olivier.cinquin@uci.edu>