fix all language links, their order, old website links and http links
This commit fixes links for all languages that point to Collections website. It also reorders them so that the order is the same as in the website's Language drop-down menu. Then it uses HTTPS instead of HTTP. Finally, it uses new eclipse.dev/collections instead of old www.eclipse.org/collections.